split app.py into logic and http handlers
