+++ app/app.py
@@ -1,23 +1,13 @@
-import contextlib
-import datetime
-import glob
 import os
-import random
-import shutil
-import string
-import time
 import urllib.request
-import uuid
 
 import filetype
+import imgpush
 import settings
-import timeout_decorator
 from flask import Flask, Response, jsonify, request, send_from_directory
 from flask_cors import CORS
 from flask_limiter import Limiter
 from flask_limiter.util import get_remote_address
-from wand.exceptions import MissingDelegateError
-from wand.image import Image
 from werkzeug.middleware.proxy_fix import ProxyFix
 
 app = Flask(__name__)
@@ -30,13 +20,6 @@ limiter = Limiter(get_remote_address, app=app, default_limits=[])
 app.use_x_sendfile = True
 
 
-if settings.NUDE_FILTER_MAX_THRESHOLD:
-    from nudenet import NudeClassifier
-    nude_classifier = NudeClassifier()
-else:
-    nude_classifier = None
-
-
 @app.after_request
 def after_request(resp):
     x_sendfile = resp.headers.get("X-Sendfile")
@@ -47,113 +30,6 @@ def after_request(resp):
     return resp
 
 
-class InvalidSizeError(Exception):
-    pass
-
-
-class CollisionError(Exception):
-    pass
-
-
-def _get_size_from_string(size):
-    try:
-        size = int(size)
-        if len(settings.VALID_SIZES) and size not in settings.VALID_SIZES:
-            raise InvalidSizeError
-    except ValueError:
-        size = ""
-    return size
-
-
-def _clear_imagemagick_temp_files():
-    """
-    A bit of a hacky solution to prevent exhausting the cache ImageMagick uses on disk.
-    It works by checking for imagemagick cache files under /tmp/
-    and removes those that are older than settings.MAX_TMP_FILE_AGE in seconds.
-    """
-    imagemagick_temp_files = glob.glob("/tmp/magick-*")
-    for filepath in imagemagick_temp_files:
-        modified = datetime.datetime.strptime(
-            time.ctime(os.path.getmtime(filepath)), "%a %b %d %H:%M:%S %Y",
-        )
-        diff = datetime.datetime.now() - modified
-        seconds = diff.seconds
-        if seconds > settings.MAX_TMP_FILE_AGE:
-            os.remove(filepath)
-
-
-def _get_random_filename() -> str:
-    random_string = _generate_random_filename()
-    if settings.NAME_STRATEGY == "randomstr":
-        file_exists = len(glob.glob(f"{settings.IMAGES_DIR}/{random_string}.*")) > 0
-        if file_exists:
-            return _get_random_filename()
-    return random_string
-
-
-def _generate_random_filename() -> str:
-    if settings.NAME_STRATEGY == "uuidv4":
-        return str(uuid.uuid4())
-    if settings.NAME_STRATEGY == "randomstr":
-        return "".join(
-            random.choices(
-                string.ascii_lowercase + string.digits + string.ascii_uppercase, k=5
-            )
-        )
-    return ""  # Add default return for type safety
-
-
-def _resize_image(path, width, height):
-    _, extension = os.path.splitext(path)
-
-    is_animated_webp = False
-
-    with Image(filename=path) as src:
-        is_animated_webp = extension == ".webp" and len(src.sequence) > 1
-
-        if is_animated_webp:
-            img = src.convert("gif")
-        else:
-            img = src.clone()
-
-    current_aspect_ratio = img.width / img.height
-
-    if not width:
-        width = int(current_aspect_ratio * height)
-
-    if not height:
-        height = int(width / current_aspect_ratio)
-
-    desired_aspect_ratio = width / height
-
-    # Crop the image to fit the desired AR
-    if desired_aspect_ratio > current_aspect_ratio:
-        newheight = int(img.width / desired_aspect_ratio)
-        img.crop(
-            0,
-            int((img.height / 2) - (newheight / 2)),
-            width=img.width,
-            height=newheight,
-        )
-    else:
-        newwidth = int(img.height * desired_aspect_ratio)
-        img.crop(
-            int((img.width / 2) - (newwidth / 2)), 0, width=newwidth, height=img.height,
-        )
-
-    @timeout_decorator.timeout(settings.RESIZE_TIMEOUT)
-    def resize(img, width, height):
-        img.sample(width, height)
-
-    with contextlib.suppress(timeout_decorator.TimeoutError):
-        resize(img, width, height)
-
-    if is_animated_webp:
-        converted = img.convert("webp")
-        img.close()
-        return converted
-
-    return img
 
 
 @app.route("/", methods=["GET"])
@@ -182,11 +58,11 @@ def liveness():
     )
 )
 def upload_image():
-    _clear_imagemagick_temp_files()
+    imgpush.clear_imagemagick_temp_files()
 
     is_svg = False
 
-    random_string = _get_random_filename()
+    random_string = imgpush.get_random_filename()
     tmp_filepath = os.path.join("/tmp", random_string)
 
     if "file" in request.files:
@@ -198,11 +74,9 @@ def upload_image():
     else:
         return jsonify(error="File is missing!"), 400
 
-    if settings.NUDE_FILTER_MAX_THRESHOLD:
-        unsafe_val = nude_classifier.classify(tmp_filepath).get(tmp_filepath, {}).get("unsafe", 0)
-        if unsafe_val >= settings.NUDE_FILTER_MAX_THRESHOLD:
-            os.remove(tmp_filepath)
-            return jsonify(error="Nudity not allowed"), 400
+    if imgpush.check_nudity_filter(tmp_filepath):
+        os.remove(tmp_filepath)
+        return jsonify(error="Nudity not allowed"), 400
 
     file_filetype = filetype.guess_extension(tmp_filepath)
     output_type = (settings.OUTPUT_TYPE or file_filetype).replace(".", "")
@@ -212,38 +86,10 @@ def upload_image():
     elif is_svg:
         output_type = "svg"
 
-    error = None
-
     output_filename = os.path.basename(tmp_filepath) + f".{output_type}"
     output_path = os.path.join(settings.IMAGES_DIR, output_filename)
 
-    try:
-        if os.path.exists(output_path):
-            raise CollisionError
-        if output_type == "mp4":
-            if settings.ALLOW_VIDEO:
-                shutil.move(tmp_filepath, output_path)
-            else:
-                error = "Invalid Filetype"
-        elif output_type == "svg":
-            shutil.move(tmp_filepath, output_path)
-        else:
-            with Image(filename=tmp_filepath) as img:
-                img.strip()
-                if output_type not in ["gif", "webp"]:
-                    # Extract first frame for non-animated formats
-                    first_frame = img.sequence[0]  # type: ignore
-                    with Image(image=first_frame) as first_frame_img, \
-                         first_frame_img.convert(output_type) as converted:
-                        converted.save(filename=output_path)
-                else:
-                    with img.convert(output_type) as converted:
-                        converted.save(filename=output_path)
-    except MissingDelegateError:
-        error = "Invalid Filetype"
-    finally:
-        if os.path.exists(tmp_filepath):
-            os.remove(tmp_filepath)
+    error = imgpush.process_image(tmp_filepath, output_path, output_type, is_svg)
 
     if error:
         return jsonify(error=error), 400
@@ -263,9 +109,9 @@ def get_image(filename):
 
     if (width or height) and (os.path.isfile(path)) and extension != ".mp4":
         try:
-            width = _get_size_from_string(width)
-            height = _get_size_from_string(height)
-        except InvalidSizeError:
+            width = imgpush.get_size_from_string(width)
+            height = imgpush.get_size_from_string(height)
+        except imgpush.InvalidSizeError:
             return (
                 jsonify(error=f"size value must be one of {settings.VALID_SIZES}"),
                 400,
@@ -277,8 +123,8 @@ def get_image(filename):
         resized_path = os.path.join(settings.CACHE_DIR, resized_filename)
 
         if not os.path.isfile(resized_path) and (width or height):
-            _clear_imagemagick_temp_files()
-            resized_image = _resize_image(path, width, height)
+            imgpush.clear_imagemagick_temp_files()
+            resized_image = imgpush.resize_image(path, width, height)
             resized_image.strip()
             resized_image.save(filename=resized_path)
             resized_image.close()
